add partial fill ix

+++ programs/elara/src/state.rs
@@ -10,7 +10,7 @@ pub struct EscrowAccount {
     pub unique_id: u64,
     pub tokens: Tokens,
     pub amount: Amount,
-    pub slippage_bps: u64,
+    pub slippage_bps: u16,
     pub fee_bps: u64,
     pub expired_at: i64,
     pub created_at: i64,
